Aplicando Presenters and Transformers

+++ app/Http/Controllers/ProjectController.php
@@ -6,6 +6,8 @@ use CodeProject\Repositories\ProjectRepository;
 use CodeProject\Services\ProjectService;
 use Illuminate\Http\Request;
 
+//use LucaDegasperi\OAuth2Server\Facades\Authorizer;
+
 class ProjectController extends Controller
 {
     /**
@@ -26,7 +28,7 @@ class ProjectController extends Controller
 
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id' => $this->getOwnerId()]);
     }
 
     public function store(Request $request)
@@ -36,16 +38,48 @@ class ProjectController extends Controller
 
     public function show($id)
     {
+        if (!$this->checkProjectPermissions($id)) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->repository->find($id);
     }
 
     public function update(Request $request, $id)
     {
+        if (!$this->checkProjectPermissions($id)) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->service->update($request->all(), $id);
     }
 
     public function destroy($id)
     {
+        if (!$this->checkProjectOwner($id)) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         $this->repository->delete($id);
+    }
+
+    private function getOwnerId()
+    {
+        return \Authorizer::getResourceOwnerId();
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        return $this->repository->isOwner($projectId, $this->getOwnerId());
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        return $this->repository->hasMember($projectId, $this->getOwnerId());
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        return $this->checkProjectOwner($projectId) || $this->checkProjectMember($projectId);
     }
 }
